src/pyproject.toml
[tool.poetry]
name = "django-api-boilerplate"
version = "0.1.0"
description = "A django API boilerplate with django-ninja fully dockerized"
authors = ["Nicolás Salas <nicolas.salas@imfd.cl>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
django-ninja = "^0.17.0"
pydantic = "^1.9.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.8.6"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
pytest = "^6.2.4"
django-extensions = "^3.1.5"
boto3 = "^1.26"
django-storages = "^1.13"

[tool.poetry.dev-dependencies]
isort = "^5.8.0"
ipython = "^7.23.1"
ipdb = "^0.13.8"
flake8 = "^3.9.2"
flake8-print = "^4.0.0"
pytest-django = "^4.3.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "conf.settings"
python_files = "tests.py test_*.py *_tests.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
